Implemented UI and a new Route for joining a random Game or creating/joining custom games
