+++ public/css/base.css
@@ -1,0 +1,81 @@
+* {
+    margin: 0;
+    padding: 0;
+    font-family: "Cascadia Code";
+}
+
+*,
+*::after,
+*::before {
+    box-sizing: border-box;
+}
+
+:root {
+    --clr-complementry: rgb(255, 210, 11);
+}
+
+ul {
+    list-style: none;
+}
+
+body {
+    width: 100vw;
+    background: #1a1a1a;
+    color: white;
+}
+
+a {
+    text-decoration: none;
+    color: white;
+    transition: 150ms ease-in-out;
+    position: relative;
+}
+
+.form-control {
+    background: transparent;
+    border: 0;
+    border: 1px solid white;
+    border-radius: 2px;
+    color: white;
+    padding: 10px;
+    min-height: 20px;
+    line-height: 20px;
+    width: 100%;
+}
+
+.form-control:hover,
+.form-control:focus {
+    outline: none;
+}
+
+.button {
+    padding: 0.75rem 1rem;
+    background-color: #3a3a3a;
+    color: white;
+    border: 0;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: all 150ms ease-in-out;
+}
+
+.button:hover,
+.button:focus {
+    outline: none;
+    background: var(--clr-complementry);
+    color: #2a2a2a;
+}
+
+.button-block {
+    width: 100%;
+}
+
+.divider {
+    width: 100px;
+    background: white;
+    height: 2px;
+}
+
+.btn-active {
+    background-color: var(--clr-complementry);
+    color: #2a2a2a;
+}
