+++ routes/index.js
@@ -3,4 +3,8 @@ const router = require("express").Router();
 router.get("/", (req, res) => {
   res.render("index");
 });
+
+router.get("/drawpad", (req, res) => {
+  res.render("drawpad");
+});
 module.exports = router;
