+++ server.js
@@ -70,7 +70,5 @@ io.on("connection", (socket) => {
   });
 });
 
-// socket
-
 const PORT = process.env.PORT || 3000;
 server.listen(PORT, () => console.log(`Server Running at PORT ${PORT}`));
