+++ views/drawpad.pug
@@ -1,0 +1,49 @@
+
+extends layouts/layouts
+
+block stylesheets 
+    link(rel="stylesheet", href="/css/drawpad.css")
+
+
+block content
+    .container 
+        .user-container 
+            ul#user_list.list
+                
+        .draw-container
+            .draw-wrapper
+                ul.menu
+                    li.menu-item.no-bg
+                        div
+                            input#pencil.d-none(type="radio", name="pencil" checked) 
+                            label(for="pencil").button.pencil Pencil
+                        div
+                            input#eraser.d-none(type="radio", name="pencil") 
+                            label(for="eraser").button.eraser Eraser
+                    li.menu-item
+                        label(for="bg_color_picker") Bg Color
+                        input#bg_color_picker(type="color", name="bg_color_picker")
+                    li.menu-item
+                        label(for="stroke_color_picker") Stroke Color
+                        input#stroke_color_picker(type="color", name="stroke_color_picker")
+                    li.menu-item
+                        label(for="stroke-size") Stroke Size (
+                            span#current_stroke_size_value 5
+                            span )
+                        input#stroke_size(type="range", min=5  max = 20 name="stroke_size" )
+                    
+                    button.button#reset-canvas Reset Canvas
+                canvas#canvas
+        .chat 
+            ul#chat-message-list.list 
+                li.item.bot_message
+                    h5.username Room Bot
+                    p.message-text Chat Here
+            hr
+            form#chat-form(action="#")
+                label(for="chat_message") Enter Your Message
+                input.chat-message#chat_message(name="chat_message" autocomplete="off") 
+                span.error
+block scripts
+    script(src="/socket.io/socket.io.js") 
+    script(type="module" src="/js/drawpad.js")
