+++ views/index.pug
@@ -1,48 +1,63 @@
-doctype html
-html(lang="en")
-    head
-        meta(charset="UTF-8")
-        meta(http-equiv="X-UA-Compatible", content="IE=edge")
-        meta(name="viewport", content="width=device-width, initial-scale=1.0")
-        title Document
-        link(rel="stylesheet", href="/css/index.css")
-    body 
-    .user-container 
-        ul#user_list.list
-            
-    .draw-container
-        .draw-wrapper
-            ul.menu
-                li.menu-item.no-bg
-                    div
-                        input#pencil.d-none(type="radio", name="pencil" checked) 
-                        label(for="pencil").button.pencil Pencil
-                    div
-                        input#eraser.d-none(type="radio", name="pencil") 
-                        label(for="eraser").button.eraser Eraser
-                li.menu-item
-                    label(for="bg_color_picker") Bg Color
-                    input#bg_color_picker(type="color", name="bg_color_picker")
-                li.menu-item
-                    label(for="stroke_color_picker") Stroke Color
-                    input#stroke_color_picker(type="color", name="stroke_color_picker")
-                li.menu-item
-                    label(for="stroke-size") Stroke Size (
-                        span#current_stroke_size_value 5
-                        span )
-                    input#stroke_size(type="range", min=5  max = 20 name="stroke_size" )
-                
-                button.button#reset-canvas Reset Canvas
-            canvas#canvas
-    .chat 
-        ul#chat-message-list.list 
-            li.item.bot_message
-                h5.username Room Bot
-                p.message-text Chat Here
-        hr
-        form#chat-form(action="#")
-            label(for="chat_message") Enter Your Message
-            input.chat-message#chat_message(name="chat_message" autocomplete="off") 
-            span.error
-script(src="/socket.io/socket.io.js") 
-script(type="module" src="/js/index.js")  
+extends layouts/layouts
+
+block stylesheets
+    link(rel="stylesheet", href="/css/index.css")
+
+
+block content 
+    header 
+        h1.logo Pictionary 
+        nav 
+            ul.nav-list 
+                li.nav-item
+                    a.nav-link(href="#") Info
+                li.nav-item
+                    a.nav-link(href="#") Source Code
+                li.nav-item
+                    a.nav-link(href="#") Developer
+    main 
+        .container
+            .btn-group 
+                button.button.button-block.btn-active#random-play-button Play Randoms 
+                button.button.button-block#custom-play-button Play Custom 
+            .form-content#random-room
+                h2 Play a Random Game
+                form.random-form#random_join_room(action="#") 
+                    .form-group 
+                        label(for="username") Enter Your Username
+                        input.form-control#username(type="text", name="")
+                    
+                    button.button.button-block#join_random_room Play with Randoms
+
+            .form-content#custom-room.dissappear
+                h2 Join Game
+                form.random-form#random_join_room(action="#") 
+                    .form-group 
+                        label(for="username") Enter Your Username
+                        input.form-control#username(type="text", name="")
+                    
+                form.random-form#random_join_room(action="#") 
+                    .form-group 
+                        label(for="username") Enter Your Username
+                        input.form-control#username(type="text", name="")
+                    
+                    button.button.button-block#join_random_room Join Game
+                h2 Create New Game
+                form.random-form#random_join_room(action="#") 
+                    .form-group 
+                        label(for="username") Enter Your Username
+                        input.form-control#username(type="text", name="")
+                    
+                form.random-form#random_join_room(action="#") 
+                    .form-group 
+                        label(for="username") Your Generated Room ID
+                        .room-id-readonly-field
+                            input.form-control#generate-random-room-id(type="text", name="" readonly value="Random Room ID")  
+                            span#generate-random-room-id-copy-button
+                                i.far.fa-clipboard
+                                span.copy-message 
+                                    | Room ID Copied!
+                    button.button.button-block#join_random_room Create Game
+
+block scripts 
+    script(src="/js/index.js")
